# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Audience interaction API for meetups: live questions, Q\u0026A board, word clouds, feedback, presentations, QR codes and admin console. Audience requests get an anonymous participant cookie automatically.",
        "title": "Meetup API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/events/{code}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Get event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.EventDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{code}/answers": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Submit answer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "answer",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{code}/presentations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "List presentations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.PresentationDTO"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{code}/presentations/{id}/download": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Download presentation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Presentation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{code}/qa": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "List public Q\u0026A",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.QADTO"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Create Q\u0026A question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "question",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.QARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{code}/qa/{id}/vote": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Vote Q\u0026A",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Q\u0026A ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{code}/qr.png": {
            "get": {
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Get event QR code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{code}/state": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Get event state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.StateDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{code}/stream": {
            "get": {
                "description": "Server-sent events stream of personalized state. Each message is `event: state` with JSON StateDTO.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Stream event state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "text/event-stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.Result": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                }
            }
        },
...
        "handlers.AnswerRequest": {
            "type": "object",
            "properties": {
                "question_id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "handlers.EventDTO": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "feedback_open": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pending_qa_count": {
                    "type": "integer"
                },
                "question_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.FeedbackDTO": {
            "type": "object",
            "properties": {
                "open": {
                    "type": "boolean"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.QuestionDTO"
                    }
                }
            }
        },
        "handlers.PresentationDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "speaker": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handlers.QADTO": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "voted": {
                    "type": "boolean"
                },
                "votes": {
                    "type": "integer"
                }
            }
        },
        "handlers.QARequest": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                }
            }
        },
        "handlers.QuestionDTO": {
            "type": "object",
            "properties": {
                "answered": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_feedback": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "my_answer": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "position": {
                    "type": "integer"
                },
                "prompt": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Result"
                    }
                },
                "show_results": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handlers.StateDTO": {
            "type": "object",
            "properties": {
                "active_question": {
                    "$ref": "#/definitions/handlers.QuestionDTO"
                },
                "event": {
                    "$ref": "#/definitions/handlers.EventDTO"
                },
                "feedback": {
                    "$ref": "#/definitions/handlers.FeedbackDTO"
                },
                "qa": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.QADTO"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Admin session cookie (meetup_session) issued by POST /api/auth/login or the OIDC flow.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}